requirements.txt
numpy==1.21.2
pandas==1.3.5
matplotlib==3.4.3
opencv_python==4.10.0.84
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.66.6
Pillow==8.3.2
